[ADD] Math for Cash
I added the math for the cash that we will be earning. While adding some classes if we want to make a breakdown of the money that the player earns during that day. Also checked to see if Bo's UI works with my code. It does. Just need to link it up with Dan's Code :thumbsup: